RIA-7911_Update_hearing_adjustments
