+++ src/functionalTest/resources/scenarios/RIA-7322-internal-hearing-requirements-updated-ada-letter.json
@@ -1,11 +1,11 @@
 {
-  "description": "RIA-7322 Internal hearing requirements updated letter - ADA",
+  "description": "RIA-7322 Internal hearing adjustments updated letter - ADA",
   "request": {
     "uri": "/asylum/ccdAboutToSubmit",
     "credentials": "CaseOfficer",
     "input": {
       "id": 7322,
-      "eventId": "updateHearingRequirements",
+      "eventId": "updateHearingAdjustments",
       "state": "awaitingRespondentEvidence",
       "caseData": {
         "template": "minimal-internal-appeal-submitted.json",
@@ -28,10 +28,10 @@
           {
             "id": "1",
             "value": {
-              "tag": "internalHearingRequirementsUpdatedLetter",
+              "tag": "internalHearingAdjustmentsUpdatedLetter",
               "document": {
                 "document_url": "$/http.+\/documents/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/",
-                "document_filename": "PA 12345 2019-Awan-hearing-requirements-updated-letter.PDF",
+                "document_filename": "PA 12345 2019-Awan-hearing-adjustments-updated-letter.PDF",
                 "document_binary_url": "$/http.+\/documents/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\/binary/"
               },
               "suppliedBy": "",
