+++ src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/domain/entities/ccd/Event.java
@@ -19,7 +19,6 @@ public enum Event {
     ADD_APPEAL_RESPONSE("addAppealResponse", CaseType.ASYLUM),
     REQUEST_HEARING_REQUIREMENTS("requestHearingRequirements", CaseType.ASYLUM),
     DRAFT_HEARING_REQUIREMENTS("draftHearingRequirements", CaseType.ASYLUM),
-    UPDATE_HEARING_REQUIREMENTS("updateHearingRequirements", CaseType.ASYLUM),
     CHANGE_DIRECTION_DUE_DATE("changeDirectionDueDate", CaseType.ASYLUM),
     UPLOAD_ADDITIONAL_EVIDENCE("uploadAdditionalEvidence", CaseType.ASYLUM),
     LIST_CASE("listCase", CaseType.ASYLUM),
@@ -70,6 +69,10 @@ public enum Event {
     MAKE_NEW_APPLICATION("makeNewApplication", CaseType.BAIL),
     EDIT_BAIL_APPLICATION_AFTER_SUBMIT("editBailApplicationAfterSubmit", CaseType.BAIL),
     UPLOAD_SIGNED_DECISION_NOTICE("uploadSignedDecisionNotice", CaseType.BAIL),
+
+    UPDATE_HEARING_REQUIREMENTS("updateHearingRequirements", CaseType.ASYLUM),
+    UPDATE_HEARING_ADJUSTMENTS("updateHearingAdjustments", CaseType.ASYLUM),
+
     @JsonEnumDefaultValue
     UNKNOWN("unknown", CaseType.UNKNOWN);
