--- src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/domain/handlers/presubmit/letter/InternalHearingRequirementsUpdatedLetterGenerator.java
@@ -1,73 +1,0 @@
-package uk.gov.hmcts.reform.iacasedocumentsapi.domain.handlers.presubmit.letter;
-
-import static java.util.Objects.requireNonNull;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCaseDefinition.NOTIFICATION_ATTACHMENT_DOCUMENTS;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.AsylumCaseUtils.isAppellantInDetention;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.AsylumCaseUtils.isInternalCase;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCase;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.DocumentTag;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.CaseDetails;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.Event;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.callback.Callback;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.callback.PreSubmitCallbackStage;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.field.Document;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.handlers.PreSubmitCallbackHandler;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.DocumentCreator;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.DocumentHandler;
-
-@Component
-public class InternalHearingRequirementsUpdatedLetterGenerator implements PreSubmitCallbackHandler<AsylumCase> {
-
-    private final DocumentCreator<AsylumCase> internalHearingRequirementsUpdatedLetter;
-    private final DocumentHandler documentHandler;
-
-    public InternalHearingRequirementsUpdatedLetterGenerator(
-        @Qualifier("internalHearingRequirementsUpdatedLetter") DocumentCreator<AsylumCase> internalHearingRequirementsUpdatedLetter,
-        DocumentHandler documentHandler
-    ) {
-        this.internalHearingRequirementsUpdatedLetter = internalHearingRequirementsUpdatedLetter;
-        this.documentHandler = documentHandler;
-    }
-
-    public boolean canHandle(
-        PreSubmitCallbackStage callbackStage,
-        Callback<AsylumCase> callback
-    ) {
-        requireNonNull(callbackStage, "callbackStage must not be null");
-        requireNonNull(callback, "callback must not be null");
-        AsylumCase asylumCase = callback.getCaseDetails().getCaseData();
-
-        return callbackStage == PreSubmitCallbackStage.ABOUT_TO_SUBMIT
-               && callback.getEvent() == Event.UPDATE_HEARING_REQUIREMENTS
-               && isInternalCase(asylumCase)
-               && isAppellantInDetention(asylumCase);
-    }
-
-    public PreSubmitCallbackResponse<AsylumCase> handle(
-        PreSubmitCallbackStage callbackStage,
-        Callback<AsylumCase> callback
-    ) {
-        if (!canHandle(callbackStage, callback)) {
-            throw new IllegalStateException("Cannot handle callback");
-        }
-
-        final CaseDetails<AsylumCase> caseDetails = callback.getCaseDetails();
-        final AsylumCase asylumCase = caseDetails.getCaseData();
-
-        Document documentForUpload = internalHearingRequirementsUpdatedLetter.create(caseDetails);
-
-        documentHandler.addWithMetadata(
-            asylumCase,
-            documentForUpload,
-            NOTIFICATION_ATTACHMENT_DOCUMENTS,
-            DocumentTag.INTERNAL_HEARING_REQUIREMENTS_UPDATED_LETTER
-        );
-
-        return new PreSubmitCallbackResponse<>(asylumCase);
-    }
-
-}
