--- src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/domain/templates/letter/InternalHearingRequirementsUpdatedLetterTemplate.java
@@ -1,51 +1,0 @@
-package uk.gov.hmcts.reform.iacasedocumentsapi.domain.templates.letter;
-
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.AsylumCaseUtils.getAppellantPersonalisation;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.DateUtils.formatDateForNotificationAttachmentDocument;
-
-import java.util.HashMap;
-import java.util.Map;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.DateProvider;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCase;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.CaseDetails;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.templates.DocumentTemplate;
-import uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.CustomerServicesProvider;
-
-@Component
-public class InternalHearingRequirementsUpdatedLetterTemplate implements DocumentTemplate<AsylumCase> {
-
-    private final String templateName;
-    private final DateProvider dateProvider;
-    private final CustomerServicesProvider customerServicesProvider;
-
-    public InternalHearingRequirementsUpdatedLetterTemplate(
-        @Value("${internalHearingRequirementsUpdated.templateName}") String templateName,
-        DateProvider dateProvider,
-        CustomerServicesProvider customerServicesProvider) {
-        this.templateName = templateName;
-        this.dateProvider = dateProvider;
-        this.customerServicesProvider = customerServicesProvider;
-    }
-
-    public String getName() {
-        return templateName;
-    }
-
-    public Map<String, Object> mapFieldValues(
-        CaseDetails<AsylumCase> caseDetails
-    ) {
-        final AsylumCase asylumCase = caseDetails.getCaseData();
-        final Map<String, Object> fieldValues = new HashMap<>();
-
-
-        fieldValues.putAll(getAppellantPersonalisation(asylumCase));
-        fieldValues.put("dateLetterSent", formatDateForNotificationAttachmentDocument(dateProvider.now()));
-        fieldValues.put("customerServicesTelephone", customerServicesProvider.getInternalCustomerServicesTelephone(asylumCase));
-        fieldValues.put("customerServicesEmail", customerServicesProvider.getInternalCustomerServicesEmail(asylumCase));
-
-        return fieldValues;
-    }
-
-}
