+++ src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/infrastructure/config/DocumentCreatorConfiguration.java
@@ -1234,13 +1234,13 @@ public class DocumentCreatorConfiguration {
         );
     }
 
-    @Bean("internalHearingRequirementsUpdatedLetter")
-    public DocumentCreator<AsylumCase> getInternalHearingRequirementsUpdatedLetterCreator(
-        @Value("${internalHearingRequirementsUpdated.contentType}") String contentType,
-        @Value("${internalHearingRequirementsUpdated.fileExtension}") String fileExtension,
-        @Value("${internalHearingRequirementsUpdated.fileName}") String fileName,
+    @Bean("internalHearingAdjustmentsUpdatedLetter")
+    public DocumentCreator<AsylumCase> getInternalHearingAdjustmentsUpdatedLetterCreator(
+        @Value("${internalHearingAdjustmentsUpdated.contentType}") String contentType,
+        @Value("${internalHearingAdjustmentsUpdated.fileExtension}") String fileExtension,
+        @Value("${internalHearingAdjustmentsUpdated.fileName}") String fileName,
         AsylumCaseFileNameQualifier fileNameQualifier,
-        InternalHearingRequirementsUpdatedLetterTemplate documentTemplate,
+        InternalHearingAdjustmentsUpdatedLetterTemplate documentTemplate,
         DocumentGenerator documentGenerator,
         DocumentUploader documentUploader
     ) {
