+++ src/test/java/uk/gov/hmcts/reform/iacasedocumentsapi/domain/templates/letter/InternalHearingAdjustmentsUpdatedLetterTemplateTest.java
@@ -1,0 +1,84 @@
+package uk.gov.hmcts.reform.iacasedocumentsapi.domain.templates.letter;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCaseDefinition.*;
+import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.DateUtils.formatDateForNotificationAttachmentDocument;
+
+import java.time.LocalDate;
+import java.util.Map;
+import java.util.Optional;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.mockito.junit.jupiter.MockitoSettings;
+import org.mockito.quality.Strictness;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.DateProvider;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCase;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.CaseDetails;
+import uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.CustomerServicesProvider;
+
+@ExtendWith(MockitoExtension.class)
+@SuppressWarnings("unchecked")
+@MockitoSettings(strictness = Strictness.LENIENT)
+public class InternalHearingAdjustmentsUpdatedLetterTemplateTest {
+
+    @Mock private CaseDetails<AsylumCase> caseDetails;
+    @Mock private AsylumCase asylumCase;
+    @Mock private CustomerServicesProvider customerServicesProvider;
+    @Mock private DateProvider dateProvider;
+    private InternalHearingAdjustmentsUpdatedLetterTemplate internalHearingAdjustmentsUpdatedLetterTemplate;
+    private final String templateName = "TB-IAC-LET-ENG-00030.docx";
+    private final String appealReferenceNumber = "HU/11111/2023";
+    private final String homeOfficeReferenceNumber = "A1234567/001";
+    private final String appellantGivenNames = "John";
+    private final String appellantFamilyName = "Doe";
+    private final String customerServicesTelephone = "0300 123 1711";
+    private final String customerServicesEmail = "email@example.com";
+
+    @BeforeEach
+    void setUp() {
+        internalHearingAdjustmentsUpdatedLetterTemplate =
+            new InternalHearingAdjustmentsUpdatedLetterTemplate(
+                templateName,
+                dateProvider,
+                customerServicesProvider
+                );
+    }
+
+    @Test
+    void should_return_template_name() {
+
+        assertEquals(templateName, internalHearingAdjustmentsUpdatedLetterTemplate.getName());
+    }
+
+    void dataSetUp() {
+        when(caseDetails.getCaseData()).thenReturn(asylumCase);
+
+        when(asylumCase.read(APPEAL_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(appealReferenceNumber));
+        when(asylumCase.read(HOME_OFFICE_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(homeOfficeReferenceNumber));
+        when(asylumCase.read(APPELLANT_GIVEN_NAMES, String.class)).thenReturn(Optional.of(appellantGivenNames));
+        when(asylumCase.read(APPELLANT_FAMILY_NAME, String.class)).thenReturn(Optional.of(appellantFamilyName));
+        when(customerServicesProvider.getInternalCustomerServicesTelephone(asylumCase)).thenReturn(customerServicesTelephone);
+        when(customerServicesProvider.getInternalCustomerServicesEmail(asylumCase)).thenReturn(customerServicesEmail);
+        when(dateProvider.now()).thenReturn(LocalDate.now());
+    }
+
+    @Test
+    void should_map_case_data_to_template_field_values() {
+        dataSetUp();
+
+        Map<String, Object> templateFieldValues = internalHearingAdjustmentsUpdatedLetterTemplate.mapFieldValues(caseDetails);
+
+        assertEquals(8, templateFieldValues.size());
+        assertEquals(appealReferenceNumber, templateFieldValues.get("appealReferenceNumber"));
+        assertEquals(homeOfficeReferenceNumber, templateFieldValues.get("homeOfficeReferenceNumber"));
+        assertEquals(appellantGivenNames, templateFieldValues.get("appellantGivenNames"));
+        assertEquals(appellantFamilyName, templateFieldValues.get("appellantFamilyName"));
+        assertEquals(formatDateForNotificationAttachmentDocument(dateProvider.now()), templateFieldValues.get("dateLetterSent"));
+        assertEquals(customerServicesTelephone, templateFieldValues.get("customerServicesTelephone"));
+        assertEquals(customerServicesEmail, templateFieldValues.get("customerServicesEmail"));
+    }
+}
